backend/requirements.txt
Django==4.2.25
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.0
psycopg[binary]==3.2.10
